# management/src/main/resources/swagger.yaml
openapi: 3.0.1
info:
  title: Security Guard Management System
  description: Managing the guards data
  version: 1.0.0
paths:
  /security:
    post:
      tags:
        - security
      summary: Add a new guard
      operationId: createGuardDetails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityDTO'
      responses:
        '201':
          description: Profile created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityDTO'

    get:
      tags:
        - security

      summary: Get all guards details

      operationId: getAllGuards

      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0

        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 5

      responses:
        '200':
          description: Successful

          content:
            application/json:
              schema:
                type: array

                items:
                  $ref: '#/components/schemas/SecurityDTO'


  /security/search:
    post:
      tags:
        - security
      summary: Search security with advanced filters
      operationId: searchSecurity
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterDTO'
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecurityDTO'


  /security/{guardId}:
    get:
      tags:
        - security
      summary: Get security by ID
      operationId: getsecurityById
      parameters:
        - name: guardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityDTO'

        '404':
          description: security not found

    put:
      tags:
        - security
      summary: Update a security details
      operationId: updateSecurity
      parameters:
        - name: guardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityDTO'
      responses:
        '200':
          description: profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityDTO'

    delete:
      tags:
        - security
      summary: Delete profile
      operationId: deleteProfile
      parameters:
        - name: guardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Profile deleted successfully
          content:
            application/json:
              schema:
                type: string
                example: "Profile deleted successfully"


components:
  schemas:
    SecurityDTO:
      type: object
      properties:
        guardId:
          type: integer
        name:
          type: string
        address:
          type: string
        age:
          type: integer
        dob:
          type: string
          format: date
        qualification:
          type: string
        shiftType:
          type: string
        experience:
          type: string
        mobile1:
          type: string
        mobile2:
          type: string
        joinedDate:
          type: string
          format: date
        status:
          type: string
          enum:
            - Active
            - InActive
        assignedLocation:
          type: string


    FilterDTO:
      type: object
      description: filter based on data
      properties:
        name:
          type: string
        guardId:
          type: integer
        shiftType:
          type: string
        qualification:
          type: string
        assignedLocation:
          type: string
        status:
          type: string
          enum:
            - Active
            - InActive